# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TodoMaster API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/task": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Get all task",
                "description": "Get all task task of the authenticated user",
                "operationId": "GetAllUserTask",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "selects the number of page in the paginated data",
                        "required": false,
                        "example": "1"
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "field name of what to sort",
                        "required": false,
                        "example": "priority"
                    },
                    {
                        "name": "sortOrder",
                        "in": "query",
                        "description": "Sort order, either 'asc' or 'desc'",
                        "required": false,
                        "example": "asc"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "List of task",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Task"
                                            }
                                        },
                                        "current_page": {
                                            "description": "The current page",
                                            "type": "string"
                                        },
                                        "next_page_url": {
                                            "description": "The url link for the next page",
                                            "type": "string"
                                        },
                                        "per_page": {
                                            "description": "The items per page",
                                            "type": "integer"
                                        },
                                        "total": {
                                            "description": "The number of total items",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Create Task",
                "description": "Create Task",
                "operationId": "TaskCreate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Task"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponseJson"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/task/{id}": {
            "put": {
                "tags": [
                    "Task"
                ],
                "summary": "Update Task",
                "description": "Update Task",
                "operationId": "UpdateTask",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the task",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Task"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "You can only update your own task"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponseJson"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Task"
                ],
                "summary": "Delete Task",
                "description": "Delete Task",
                "operationId": "DeleteTask",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the task",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Register Successfully"
                    },
                    "401": {
                        "description": "You can only delete your own task"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ValidationErrorResponseJson": {
                "properties": {
                    "message": {
                        "description": "Error message",
                        "type": "string",
                        "example": "The status field is required. (and 1 more error)",
                        "nullable": "false"
                    },
                    "errors": {
                        "description": "Error object",
                        "properties": {
                            "status": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": "['The status field is required.']",
                                "nullable": false
                            },
                            "priority": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": "['The priority field is required.']",
                                "nullable": false
                            }
                        },
                        "type": "object",
                        "nullable": "true"
                    }
                },
                "type": "object"
            },
...
            "Task": {
                "properties": {
                    "id": {
                        "description": "Task identifier",
                        "type": "integer",
                        "example": "1",
                        "nullable": "false"
                    },
                    "title": {
                        "description": "Title of the task",
                        "type": "string",
                        "example": "Create Swagger Documentation",
                        "nullable": "false"
                    },
                    "description": {
                        "description": "Destription of the Task",
                        "type": "string",
                        "example": "Lorem ipsum dolor",
                        "nullable": "true"
                    },
                    "due_date": {
                        "description": "Due date of the Task with timezone",
                        "type": "string",
                        "example": "2004-10-19 16:23:54.000 +0800",
                        "nullable": "true"
                    },
                    "status": {
                        "description": "Status of the Task if it is completed",
                        "type": "integer",
                        "example": 1,
                        "nullable": "true"
                    },
                    "priority": {
                        "description": "Priority of the Task",
                        "type": "integer",
                        "example": "1",
                        "nullable": "true"
                    },
                    "is_archived": {
                        "description": "Flag to archive a task",
                        "type": "boolean",
                        "example": true,
                        "nullable": "true"
                    },
                    "order": {
                        "description": "Sort order of a task",
                        "type": "integer",
                        "example": 1,
                        "nullable": "true"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}